+++ src/routes/graphql/graphqlSchema.ts
@@ -1,6 +1,6 @@
 import { GraphQLSchema } from 'graphql';
 import { memberTypeType, postType, profileType, userType } from './types.js';
-import { queryType } from './queries.js';
+import { queryType } from './queries/queryTypes.js';
 import { mutationType } from './mutations/mutations.js';
 
 export const RootSchema: GraphQLSchema = new GraphQLSchema({
